Added module 'memoryinspect'
When compiling with -DDEBUG, every call of malloc and free will be
monitored to see if there is any memory leak. Modules that use memory
allocation have to change include <stdlib.h> to include
"memoryinspect.h". If DEBUG definition isn't set, then the behavior of
memory allocation is left unchanged.