Few changes to allocate static data for classes
Each field now has an offset that can be used to access its value in an
array of bytes that hold the class static data, or the attributes in
case it's an instance of the class. Resolution of classes also has
another parameter to enable retrieval of the class if it has already
been loaded.